serve / watch circularity
