+++ src/command/serve/types.ts
@@ -1,0 +1,26 @@
+/*
+* types.ts
+*
+* Copyright (C) 2020 by RStudio, PBC
+*
+*/
+
+import { ServerRequest } from "http/server.ts";
+
+import { ProjectContext } from "../../project/types.ts";
+
+export interface ProjectWatcher {
+  handle: (req: ServerRequest) => boolean;
+  connect: (req: ServerRequest) => Promise<void>;
+  injectClient: (file: Uint8Array) => Uint8Array;
+  serveProject: () => ProjectContext;
+  refreshProject: () => Promise<ProjectContext>;
+}
+
+export type ServeOptions = {
+  port: number;
+  render: string;
+  browse?: boolean;
+  watch?: boolean;
+  navigate?: boolean;
+};
